moved tests to test folder
