+++ back-end/test/test-30day-meal-plan.js
@@ -1,11 +1,11 @@
 // Test script specifically for 30-day meal plan generation
 require('dotenv').config();
 const mongoose = require('mongoose');
-const connectDB = require('./config/db');
-const User = require('./models/User');
-const BiometricData = require('./models/BiometricData');
-const MealPlan = require('./models/MealPlan');
-const { generateMealPlan, getMealPlan } = require('./controllers/mealPlanController');
+const connectDB = require('../config/db');
+const User = require('../models/User');
+const BiometricData = require('../models/BiometricData');
+const MealPlan = require('../models/MealPlan');
+const { generateMealPlan, getMealPlan } = require('../controllers/mealPlanController');
 
 async function test30DayMealPlan() {
   try {
